requirements.txt
appdirs==1.4.4
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
python-dotenv==0.18.0
regex==2021.7.6
requests==2.25.1
six==1.16.0
spotipy==2.18.0
toml==0.10.2
urllib3==1.26.6
